spec-kit-rag-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
openai-agents==0.28.0
cohere==4.4.1
qdrant-client==1.7.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
markdown==3.5
PyYAML==6.0.1